requirements.txt
django==5.2.5
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
structlog==24.4.0
pytest==8.3.3
pytest-django==4.8.0
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.1
openai==1.43.0
httpx==0.27.2
certifi==2024.7.4
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
dateparser==1.2.0

